pyproject.toml
[project]
name = "homelab"
version = "1.0.0"
description = "Scripts for managing the homelab"
readme = "README.md"
dependencies = [
    "ansible>=10.3.0",
    "pyexiftool>=0.5.6",
    "python-dateutil>=2.9.0.post0",
]
requires-python = ">=3.10"

[tool.uv]
dev-dependencies = [
    "isort>=5.13.2",
    "ruff>=0.6.4",
]
